Homework Instructions
Given a monthly deposit, expressed in dollars, we want to compute the compound value after 6 months. We will base our computations on a 5 percent annual interest rate which corresponds to a 0.05 / 12 = 0.00417 monthly interest rate. 
We start with deposit dollars in our account. The value after the first month will be given by the following formula;
balance = deposit * (1 + 0.00417)
We may now update our balance again for the 2nd month as follows;
balance = (balance + deposit) * (1 + 0.05 / 12)
The same formula may be applied repeatedly to our balance in order to get the balance for the next months. 
Write a program which reads from the console a double value for the monthly deposit, then displays the balance for each months. Please note that we will not use a loop since this module did not introduce them yet; instead simply apply the formula several times.